Allow -brv or -brlow instead of -s.
Add -brv <targetbitrate_in_kbps>, which allows the user to select a bitrate manually.
Using -brlow <percentage> without selecting a target size will lower the input video's video bitrate and set it as the target. Not to be confused with lowering the video's file size by a percentage, though in practice this might achieve a similar result.